moved evenstate to its own package
